requirements.txt
fastapi
pathlib
starlette
python-dotenv
datetime
psycopg2
matplotlib
uvicorn
jinja2
httpx